# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K_WALAH Emotional AI Spotify DJ API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SpotifyUser {
  id: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface CacheStatus {
  synced: boolean;
  totalTracks: number;
  /** @nullable */
  lastSyncedAt: string | null;
  isSyncing: boolean;
  /** @nullable */
  syncProgress?: number | null;
  /** @nullable */
  syncTotal?: number | null;
}

export interface SyncResponse {
  message: string;
  started: boolean;
}

export type PlaylistRequestMode = typeof PlaylistRequestMode[keyof typeof PlaylistRequestMode];


export const PlaylistRequestMode = {
  strict: 'strict',
  balanced: 'balanced',
  chaotic: 'chaotic',
} as const;

export interface PlaylistRequest {
  /** @minLength 1 */
  vibe: string;
  mode?: PlaylistRequestMode;
  /**
     * @minimum 10
     * @maximum 100
     */
  length?: number;
}

export interface EmotionProfile {
  energy: number;
  valence: number;
  tension: number;
  nostalgia: number;
  calm: number;
  /** @nullable */
  environment?: string | null;
  /** @nullable */
  timeOfDay?: string | null;
  /** @nullable */
  motionState?: string | null;
}

export interface PlaylistTrack {
  id: string;
  name: string;
  artist: string;
  album: string;
  /** @nullable */
  albumArt?: string | null;
  durationMs?: number;
  /** @nullable */
  energy?: number | null;
  /** @nullable */
  valence?: number | null;
  /** @nullable */
  tempo?: number | null;
  score?: number;
}

export interface PlaylistResult {
  playlistId: string;
  playlistUrl: string;
  name: string;
  vibe: string;
  mode: string;
  tracks: PlaylistTrack[];
  emotionProfile: EmotionProfile;
  totalTracks?: number;
}

export interface PlaylistHistoryItem {
  id: number;
  playlistId: string;
  playlistUrl: string;
  name: string;
  vibe: string;
  mode: string;
  trackCount: number;
  createdAt: string;
  emotionProfile?: EmotionProfile;
}

export type AuthCallbackParams = {
code?: string;
error?: string;
};