# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: K_WALAH Emotional AI Spotify DJ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/callback:
    get:
      operationId: authCallback
      tags: [auth]
      summary: Spotify OAuth callback
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Redirect to app
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotifyUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SpotifyUser:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
...