# lib/api-zod/src/generated/types/spotifyUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K_WALAH Emotional AI Spotify DJ API
 * OpenAPI spec version: 0.1.0
 */

export interface SpotifyUser {
  id: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  country?: string | null;
}